# packages/contracts/generated/spring/src/main/java/tv/lumo/api/generated/api/SourcesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package tv.lumo.api.generated.api;

import tv.lumo.api.generated.model.CreateSourceRequest;
import tv.lumo.api.generated.model.Problem;
import tv.lumo.api.generated.model.Source;
import tv.lumo.api.generated.model.SourceList;
import java.util.UUID;
import tv.lumo.api.generated.model.UpdateSourceRequest;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.14.0")
@Validated
public interface SourcesApi {

    /**
     * POST /sources : Register a source, then ingest it asynchronously
     * Two distinct error surfaces, deliberately:  1. **Synchronous validation** — reachability and credentials are checked    before answering, within a bounded timeout. A failure here is a &#x60;422&#x60;    carrying an &#x60;IngestionErrorCode&#x60; and **no source is created**. This is    what makes \&quot;your credentials were refused by the server\&quot; appear while    the user is still looking at the form (US-06). 2. **Asynchronous ingestion** — once validation passes the response is    &#x60;202&#x60; with &#x60;status: PENDING&#x60;. Catalogue parsing runs in the    background. A failure there is *not* an HTTP error: the source moves    to &#x60;status: ERROR&#x60; and carries the reason in &#x60;error_code&#x60;, which the    client reads while polling &#x60;GET /sources/{id}&#x60;.  An Xtream &#x60;host&#x60; is normalised server-side: with or without a scheme, with or without a port, with or without a trailing slash. Tolerate and normalise rather than reject.  &#x60;kind: M3U_FILE&#x60; cannot be created through this JSON endpoint — see the note on &#x60;SourceKind&#x60;. 
     *
     * @param createSourceRequest  (required)
     * @return Validated and accepted. Ingestion runs in the background; the source is returned in &#x60;PENDING&#x60;. Poll &#x60;GET /sources/{id}&#x60; until &#x60;READY&#x60; or &#x60;ERROR&#x60;.  (status code 202)
     *         or The request is malformed or fails validation (&#x60;VALIDATION_FAILED&#x60;). (status code 400)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or Synchronous validation of the source failed. &#x60;code&#x60; is one of the &#x60;IngestionErrorCode&#x60; values and is what the client translates into an actionable message.  This is the number-one friction point of onboarding. \&quot;Something went wrong\&quot; loses users here; \&quot;your credentials were refused by the server\&quot; recovers them.  (status code 422)
     *         or Rate limit exceeded (&#x60;RATE_LIMITED&#x60;). (status code 429)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/sources",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Source> createSource(
         @Valid @RequestBody CreateSourceRequest createSourceRequest
    );


    /**
     * DELETE /sources/{id} : Delete a source and everything ingested from it
     * Cascades: categories, channels, EPG programmes and favourites belonging to this source are removed with it. Irreversible. 
     *
     * @param id Resource identifier. (required)
     * @return Source and its ingested data deleted. (status code 204)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or No such source on this account (&#x60;SOURCE_NOT_FOUND&#x60;). (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/sources/{id}",
        produces = { "application/problem+json" }
    )
    
    ResponseEntity<Void> deleteSource(
         @PathVariable("id") UUID id
    );


    /**
     * GET /sources/{id} : One source, with its current ingestion status
     * The endpoint clients poll after &#x60;POST /sources&#x60; or &#x60;POST /sources/{id}/sync&#x60;. &#x60;status&#x60; and &#x60;error_code&#x60; carry the outcome of the background ingestion. 
     *
     * @param id Resource identifier. (required)
     * @return The source. (status code 200)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or No such source on this account (&#x60;SOURCE_NOT_FOUND&#x60;). (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/sources/{id}",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<Source> getSource(
         @PathVariable("id") UUID id
    );


    /**
     * GET /sources : The user&#39;s sources
     *
     * @return Every source owned by the caller. (status code 200)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/sources",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<SourceList> listSources(
        
    );


    /**
     * POST /sources/{id}/sync : Force a re-synchronisation
     *
     * @param id Resource identifier. (required)
     * @return Accepted. The source is returned in &#x60;SYNCING&#x60;. (status code 202)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or No such source on this account (&#x60;SOURCE_NOT_FOUND&#x60;). (status code 404)
     *         or A synchronisation is already running for this source (&#x60;SOURCE_SYNC_IN_PROGRESS&#x60;). (status code 409)
     *         or Too many manual synchronisations (&#x60;SOURCE_SYNC_RATE_LIMITED&#x60;).  This limit protects the *user&#39;s own* IPTV server: hammering a third-party panel gets their account throttled or banned. It is a product safeguard, not a capacity one (ADR 0005).  (status code 429)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/sources/{id}/sync",
        produces = { "application/json", "application/problem+json" }
    )
    
    ResponseEntity<Source> syncSource(
         @PathVariable("id") UUID id
    );


    /**
     * PATCH /sources/{id} : Update a source
     * Omitted properties are left unchanged. Changing any property that affects ingestion (&#x60;host&#x60;, &#x60;username&#x60;, &#x60;password&#x60;, &#x60;m3u_url&#x60;, &#x60;epg_url&#x60;) moves the source back to &#x60;PENDING&#x60; and triggers a fresh ingestion.  &#x60;password&#x60; is write-only. It is re-encrypted with AES-256-GCM and is never echoed back. 
     *
     * @param id Resource identifier. (required)
     * @param updateSourceRequest  (required)
     * @return The updated source. (status code 200)
     *         or The request is malformed or fails validation (&#x60;VALIDATION_FAILED&#x60;). (status code 400)
     *         or Missing, malformed or expired access token (&#x60;UNAUTHENTICATED&#x60;, &#x60;ACCESS_TOKEN_EXPIRED&#x60;). On &#x60;ACCESS_TOKEN_EXPIRED&#x60; the client refreshes once and replays the request.  (status code 401)
     *         or No such source on this account (&#x60;SOURCE_NOT_FOUND&#x60;). (status code 404)
     *         or Synchronous validation of the source failed. &#x60;code&#x60; is one of the &#x60;IngestionErrorCode&#x60; values and is what the client translates into an actionable message.  This is the number-one friction point of onboarding. \&quot;Something went wrong\&quot; loses users here; \&quot;your credentials were refused by the server\&quot; recovers them.  (status code 422)
     */
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/sources/{id}",
        produces = { "application/json", "application/problem+json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<Source> updateSource(
         @PathVariable("id") UUID id,
         @Valid @RequestBody UpdateSourceRequest updateSourceRequest
    );

}