# packages/contracts/openapi.yaml
openapi: 3.0.3
...
info:
  title: Lumo TV API
  version: 1.0.0
  description: |
    Backend for the Lumo TV IPTV player (`lumo-api`).

    Lumo hosts, provides and resells **no content**. The user brings their own
    source — an M3U playlist or an Xtream Codes account. This API manipulates
    metadata only: catalogue, guide, favourites, playback progress, access
    rights. **Media never transits through this infrastructure**: the player
    opens the stream directly against the user's own IPTV server.

    ## Authentication

    All operations require `Authorization: Bearer <access_token>` unless the
    operation declares `security: []`. Access tokens are short-lived JWTs
    (15 minutes). Refresh tokens are opaque, rotated on every use, and bound to
    a `device`. Presenting an already-consumed refresh token revokes the whole
    token chain for that device — that is the theft-detection mechanism.

    ## Errors

    Every error response is RFC 7807 `application/problem+json` and carries a
    stable, machine-readable `code`. Clients switch on `code` and render their
    own localised message (FR/EN). `detail` is meant for logs and is never
    displayed raw to a user.

    ## Asynchronous ingestion

    Parsing is always server-side. `POST /sources` performs a fast synchronous
    reachability and credential check, then returns `202` with the source in
    `PENDING`; the full catalogue ingestion runs in the background. Clients poll
...
    blocks an HTTP request.
  contact:
...
servers:
...
security:
  - bearerAuth: []

# ==============================================================================
# PATHS
# ==============================================================================
paths:
...
  /sources:
    get:
      tags: [sources]
      operationId: listSources
      summary: The user's sources
      responses:
        '200':
          description: Every source owned by the caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [sources]
      operationId: createSource
      summary: Register a source, then ingest it asynchronously
      description: |
        Two distinct error surfaces, deliberately:

        1. **Synchronous validation** — reachability and credentials are checked
           before answering, within a bounded timeout. A failure here is a `422`
           carrying an `IngestionErrorCode` and **no source is created**. This is
           what makes "your credentials were refused by the server" appear while
           the user is still looking at the form (US-06).
        2. **Asynchronous ingestion** — once validation passes the response is
           `202` with `status: PENDING`. Catalogue parsing runs in the
           background. A failure there is *not* an HTTP error: the source moves
           to `status: ERROR` and carries the reason in `error_code`, which the
           client reads while polling `GET /sources/{id}`.

        An Xtream `host` is normalised server-side: with or without a scheme,
        with or without a port, with or without a trailing slash. Tolerate and
        normalise rather than reject.

        `kind: M3U_FILE` cannot be created through this JSON endpoint — see the
        note on `SourceKind`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        '202':
          description: |
            Validated and accepted. Ingestion runs in the background; the source
            is returned in `PENDING`. Poll `GET /sources/{id}` until `READY` or
            `ERROR`.
          headers:
            Location:
              description: URL of the created source.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/IngestionFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /sources/{id}:
    parameters:
      - $ref: '#/components/parameters/PathId'
    get:
      tags: [sources]
      operationId: getSource
      summary: One source, with its current ingestion status
      description: |
        The endpoint clients poll after `POST /sources` or
        `POST /sources/{id}/sync`. `status` and `error_code` carry the outcome of
        the background ingestion.
      responses:
        '200':
          description: The source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/SourceNotFound' }
    patch:
      tags: [sources]
      operationId: updateSource
      summary: Update a source
      description: |
        Omitted properties are left unchanged. Changing any property that affects
        ingestion (`host`, `username`, `password`, `m3u_url`, `epg_url`) moves the
        source back to `PENDING` and triggers a fresh ingestion.

        `password` is write-only. It is re-encrypted with AES-256-GCM and is
        never echoed back.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSourceRequest'
      responses:
        '200':
          description: The updated source.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/SourceNotFound' }
        '422': { $ref: '#/components/responses/IngestionFailed' }
    delete:
      tags: [sources]
      operationId: deleteSource
      summary: Delete a source and everything ingested from it
      description: |
        Cascades: categories, channels, EPG programmes and favourites belonging
        to this source are removed with it. Irreversible.
      responses:
        '204':
          description: Source and its ingested data deleted.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/SourceNotFound' }
  /sources/{id}/sync:
    post:
      tags: [sources]
      operationId: syncSource
      summary: Force a re-synchronisation
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '202':
          description: Accepted. The source is returned in `SYNCING`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/SourceNotFound' }
        '409':
          description: A synchronisation is already running for this source (`SOURCE_SYNC_IN_PROGRESS`).
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429':
          description: |
            Too many manual synchronisations (`SOURCE_SYNC_RATE_LIMITED`).

            This limit protects the *user's own* IPTV server: hammering a
            third-party panel gets their account throttled or banned. It is a
            product safeguard, not a capacity one (ADR 0005).
          headers:
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ------------------------------------------------------------- catalog ------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived access token (15 minutes) obtained from `/auth/login`,
        `/auth/register`, `/auth/oauth/google`, `/auth/refresh` or
        `/auth/device/token`.

  # ---------------------------------------------------------- parameters ------
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Resource identifier.
      schema:
        type: string
        format: uuid
...
  headers:
    RetryAfter:
      description: Seconds to wait before retrying.
      schema:
        type: integer
        format: int32
        minimum: 1

  # ----------------------------------------------------- shared responses -----
  responses:
    BadRequest:
      description: The request is malformed or fails validation (`VALIDATION_FAILED`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: |
        Missing, malformed or expired access token (`UNAUTHENTICATED`,
        `ACCESS_TOKEN_EXPIRED`). On `ACCESS_TOKEN_EXPIRED` the client refreshes
        once and replays the request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TooManyRequests:
      description: Rate limit exceeded (`RATE_LIMITED`).
      headers:
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    SourceNotFound:
      description: No such source on this account (`SOURCE_NOT_FOUND`).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    IngestionFailed:
      description: |
        Synchronous validation of the source failed. `code` is one of the
        `IngestionErrorCode` values and is what the client translates into an
        actionable message.

        This is the number-one friction point of onboarding. "Something went
        wrong" loses users here; "your credentials were refused by the server"
        recovers them.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }

  # -------------------------------------------------------------- schemas -----
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 7807 problem detail. Every error response in this API uses this
        shape with `Content-Type: application/problem+json`.

        Clients switch on `code`, never on `title` or `detail`. `detail` is
        written for logs and support, is not localised, and is never displayed
        raw to a user.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
          description: Stable URI identifying the problem type.
          example: https://lumo.tv/errors/source-auth-failed
        title:
          type: string
          description: Short, human-readable, English, non-localised summary.
          example: Source authentication failed
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: HTTP status code, repeated for convenience.
          example: 422
        code:
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
          description: |
            Stable machine-readable code. **This is the only field a client is
            allowed to branch on.** Clients must handle an unrecognised value
            gracefully rather than crash: new codes may be added within v1.
        detail:
          type: string
          nullable: true
          description: |
            Diagnostic explanation for logs and support. Never rendered raw in
            the UI. Never contains a stream URL, an Xtream password or a token.
          example: Connection timed out after 10s
        instance:
          type: string
          format: uri-reference
          nullable: true
          description: Path of the request that produced this problem.
          example: /v1/sources
        errors:
          type: array
          nullable: true
          description: Per-field details, present on validation failures only.
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      description: One field-level validation failure.
      required: [field, code]
      properties:
        field:
          type: string
          description: JSON pointer to the offending property, relative to the request body.
          example: /password
        code:
          type: string
          description: Stable machine-readable reason.
          example: TOO_SHORT
        detail:
          type: string
          nullable: true
          description: Diagnostic explanation. Not localised, not for display.
    IngestionErrorCode:
      type: string
      description: |
        **Complete enumeration of ingestion error codes for v1.**

        These are the values that reach a user during onboarding, whether
        synchronously in a `422` `Problem.code` or asynchronously in
        `Source.error_code`. Each one must have an actionable FR and EN message
        in every client — the whole point of this enumeration is that no client
        ever falls back to a generic failure message here.

        | Value | Cause | What the user should be told |
        |---|---|---|
        | `SOURCE_UNREACHABLE` | The host did not answer, DNS failed, or the request timed out. | The server could not be reached. Offer "Try again". Must read as clearly different from a credential refusal. |
        | `SOURCE_AUTH_FAILED` | The panel rejected `username`/`password`. | The credentials were refused by the server. Keep the host in the form so only what is wrong is retyped. |
        | `SOURCE_EXPIRED` | The Xtream account itself has expired. | The subscription with the provider has expired; Lumo cannot renew it. |
        | `SOURCE_MAX_CONNECTIONS` | The panel reports the simultaneous-connection limit reached. | The subscription limits simultaneous streams; close another stream. |
        | `SOURCE_INVALID_FORMAT` | The response is not a valid M3U playlist, XMLTV guide or Xtream payload. | The URL does not return a playlist. Explain what a correct M3U URL looks like. |
        | `SOURCE_EMPTY` | Parsing succeeded but yielded zero channels. | The playlist is empty. Never show a success screen over an empty list. |
        | `SOURCE_TOO_LARGE` | The payload exceeds the ingestion size ceiling. | The playlist is too large to import. |
      enum:
        - SOURCE_UNREACHABLE
        - SOURCE_AUTH_FAILED
        - SOURCE_EXPIRED
        - SOURCE_MAX_CONNECTIONS
        - SOURCE_INVALID_FORMAT
        - SOURCE_EMPTY
        - SOURCE_TOO_LARGE
    ErrorCode:
      type: string
      description: |
        Every stable error code this API emits, ingestion codes included.

        Grouped by concern:

        - **Generic** — `VALIDATION_FAILED`, `UNAUTHENTICATED`,
          `ACCESS_TOKEN_EXPIRED`, `FORBIDDEN`, `NOT_FOUND`, `CONFLICT`,
          `RATE_LIMITED`, `INTERNAL_ERROR`.
        - **Account and sign-in** — `EMAIL_ALREADY_REGISTERED`,
          `INVALID_CREDENTIALS`, `PASSWORD_TOO_WEAK`, `OAUTH_TOKEN_INVALID`,
          `VERIFICATION_TOKEN_INVALID`, `VERIFICATION_TOKEN_EXPIRED`,
          `RESET_TOKEN_INVALID`, `RESET_TOKEN_EXPIRED`.
        - **Sessions** — `REFRESH_TOKEN_INVALID`, `REFRESH_TOKEN_REUSED`,
          `DEVICE_NOT_FOUND`.
        - **TV activation (RFC 8628)** — `AUTHORIZATION_PENDING`, `SLOW_DOWN`,
          `ACCESS_DENIED`, `EXPIRED_TOKEN`, `DEVICE_CODE_NOT_FOUND`,
          `DEVICE_CODE_EXPIRED`, `DEVICE_CODE_ALREADY_USED`.
        - **Sources and ingestion** — `SOURCE_NOT_FOUND`, `SOURCE_NOT_READY`,
          `SOURCE_SYNC_IN_PROGRESS`, `SOURCE_SYNC_RATE_LIMITED`, plus every
          `IngestionErrorCode`.
        - **Catalogue and user data** — `CHANNEL_NOT_FOUND`,
          `FAVORITE_NOT_FOUND`, `FAVORITE_ALREADY_EXISTS`,
          `FAVORITE_GROUP_NOT_FOUND`, `FAVORITE_GROUP_ALREADY_EXISTS`.
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - ACCESS_TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - RATE_LIMITED
        - INTERNAL_ERROR
        - EMAIL_ALREADY_REGISTERED
        - INVALID_CREDENTIALS
        - PASSWORD_TOO_WEAK
        - OAUTH_TOKEN_INVALID
        - VERIFICATION_TOKEN_INVALID
        - VERIFICATION_TOKEN_EXPIRED
        - RESET_TOKEN_INVALID
        - RESET_TOKEN_EXPIRED
        - REFRESH_TOKEN_INVALID
        - REFRESH_TOKEN_REUSED
        - DEVICE_NOT_FOUND
        - AUTHORIZATION_PENDING
        - SLOW_DOWN
        - ACCESS_DENIED
        - EXPIRED_TOKEN
        - DEVICE_CODE_NOT_FOUND
        - DEVICE_CODE_EXPIRED
        - DEVICE_CODE_ALREADY_USED
        - SOURCE_NOT_FOUND
        - SOURCE_NOT_READY
        - SOURCE_SYNC_IN_PROGRESS
        - SOURCE_SYNC_RATE_LIMITED
        - SOURCE_UNREACHABLE
        - SOURCE_AUTH_FAILED
        - SOURCE_EXPIRED
        - SOURCE_MAX_CONNECTIONS
        - SOURCE_INVALID_FORMAT
        - SOURCE_EMPTY
        - SOURCE_TOO_LARGE
        - CHANNEL_NOT_FOUND
        - FAVORITE_NOT_FOUND
        - FAVORITE_ALREADY_EXISTS
        - FAVORITE_GROUP_NOT_FOUND
        - FAVORITE_GROUP_ALREADY_EXISTS

    # ---- enums -------------------------------------------------------------
...
    SourceKind:
      type: string
      description: |
        What the user registered.

        `M3U_FILE` designates a playlist uploaded as a file. It exists in the
        domain model and may be returned by read operations, but v1 exposes no
        endpoint that creates one: `POST /sources` is JSON-only and accepts
        `M3U_URL` and `XTREAM`. A multipart upload operation is not part of this
        contract.
      enum: [M3U_URL, M3U_FILE, XTREAM]
    SourceStatus:
      type: string
      description: |
        Ingestion lifecycle.

        `PENDING` → accepted, not started. `SYNCING` → ingestion running.
        `READY` → catalogue usable. `ERROR` → ingestion failed, see `error_code`.

        Clients poll until `READY` or `ERROR`; those are the only terminal states.
      enum: [PENDING, SYNCING, READY, ERROR]
...
    Source:
      type: object
      description: |
        A registered source.

        **This schema has no password property and never will.** The Xtream
        password is encrypted at rest (AES-256-GCM, data key wrapped by an
        out-of-database master key) and never leaves the `source` layer — not
        even for its owner (`docs/domain-model.md` §2).
      required: [id, label, kind, status]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
          description: Name the user gave this source.
        kind:
          $ref: '#/components/schemas/SourceKind'
        host:
          type: string
          nullable: true
          description: |
            Xtream base URL, normalised server-side. Returned so an edit form can
            keep it while the user corrects only what was wrong (US-06).
          example: http://panel.example.org:8080
        username:
          type: string
          nullable: true
          description: Xtream username. The matching password is never returned.
        m3u_url:
          type: string
          nullable: true
          description: Playlist URL, for `M3U_URL` sources.
        epg_url:
          type: string
          nullable: true
          description: Optional XMLTV guide URL, independent of the playlist.
        status:
          $ref: '#/components/schemas/SourceStatus'
        error_code:
          type: string
          nullable: true
          allOf:
            - $ref: '#/components/schemas/IngestionErrorCode'
          description: |
            Why the last ingestion failed. Non-null only when `status` is
            `ERROR`. A stable code, never a free-form message — the client owns
            the wording, in FR and EN.
        last_synced_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Expiry of the user's Xtream account, as reported by the panel. Shown
            after a successful registration (US-06). Null for M3U sources.
        max_connections:
          type: integer
          format: int32
          nullable: true
          description: |
            Simultaneous streams the user's subscription allows, as reported by
            the panel. Null when unknown or not applicable.
        channel_count:
          type: integer
          format: int32
          nullable: true
          description: |
            Channels ingested from this source. Derived, not stored on the
            entity. Null until the first successful ingestion; it is what "we
            found N channels" is rendered from (US-06, US-07).
    SourceList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Source'
    CreateSourceRequest:
      type: object
      description: |
        `kind` decides which properties are required:

        | `kind` | Required | Optional |
        |---|---|---|
        | `XTREAM` | `label`, `host`, `username`, `password` | `epg_url` |
        | `M3U_URL` | `label`, `m3u_url` | `epg_url` |

        `M3U_FILE` is not creatable through this operation.

        The combination is validated server-side; a mismatch is `400`
        `VALIDATION_FAILED` with per-field `errors`.
      required: [label, kind]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 100
        kind:
          $ref: '#/components/schemas/SourceKind'
        host:
          type: string
          nullable: true
          maxLength: 500
          description: |
            Xtream base URL. Accepted with or without a scheme, with or without a
            port, with or without a trailing slash — normalised rather than
            rejected (US-06).
        username:
          type: string
          nullable: true
          maxLength: 200
        password:
          type: string
          format: password
          nullable: true
          maxLength: 200
          writeOnly: true
          description: |
            Xtream password. **Write-only, by contract.** Encrypted with
            AES-256-GCM before persistence, never returned by any operation,
            never written to a log at any level including `DEBUG`.
        m3u_url:
          type: string
          nullable: true
          maxLength: 2000
        epg_url:
          type: string
          nullable: true
          maxLength: 2000
          description: Optional XMLTV URL, a separate field from the playlist URL.
    UpdateSourceRequest:
      type: object
      description: |
        Omitted properties are left unchanged. Changing `host`, `username`,
        `password`, `m3u_url` or `epg_url` re-triggers ingestion.
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 100
        host:
          type: string
          nullable: true
          maxLength: 500
        username:
          type: string
          nullable: true
          maxLength: 200
        password:
          type: string
          format: password
          nullable: true
          maxLength: 200
          writeOnly: true
          description: Write-only. Re-encrypted on write, never returned.
        m3u_url:
          type: string
          nullable: true
          maxLength: 2000
        epg_url:
          type: string
          nullable: true
          maxLength: 2000

    # ---- catalog -----------------------------------------------------------
...